requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
mlflow==2.7.1
pytest==7.4.2
requests==2.31.0
python-binance==1.0.19
prometheus-client==0.17.1
pymongo==4.5.0
dnspython==2.4.2